feat: reverting OneLoginLocalAuthManager for preference of EnrolmentCoordinator
